+++ app/Http/Controllers/incomeController.php
@@ -167,33 +167,46 @@ class incomeController extends Controller
         }
     }
     public function delete_income($id)
-    {
+    {   
 
-        $studentinfo = DB::table('student_collection')
+        $collection_std_check = DB::table('income_info')
+                                ->where('id',$id)
+                                ->first();
+
+        if($collection_std_check->std_collection_id == 1000)
+        {
+           $studentinfo = DB::table('student_collection')
                        ->where('income_id',$id)
                        ->first();
 
-        $get_data = DB::table('student_info')
-                        ->where('id',$studentinfo->student_id)
-                        ->first();
-        $paid_ammount = $get_data->paid;
-        $due_ammount = $get_data->due;
+            $get_data = DB::table('student_info')
+                            ->where('id',$studentinfo->student_id)
+                            ->first();
+            $paid_ammount = $get_data->paid;
+            $due_ammount = $get_data->due;
 
-        $get_presentammount = DB::table('student_collection')
-                              ->where('id',$studentinfo->id)
-                              ->first();
+            $get_presentammount = DB::table('student_collection')
+                                  ->where('id',$studentinfo->id)
+                                  ->first();
 
-        $collection_ammount = $get_presentammount->collection_ammount;
+            $collection_ammount = $get_presentammount->collection_ammount;
 
-        // return $collection_ammount;
+            // return $collection_ammount;
 
-        $newPaid_ammount = $paid_ammount - $collection_ammount;
+            $newPaid_ammount = $paid_ammount - $collection_ammount;
 
-        $newDue_ammount = $due_ammount + $collection_ammount;
+            $newDue_ammount = $due_ammount + $collection_ammount;
 
-        $setdata = DB::table('student_info')->where('id',$studentinfo->student_id)->update(['paid'=>$newPaid_ammount,'due'=>$newDue_ammount]);
+            $setdata = DB::table('student_info')->where('id',$studentinfo->student_id)->update(['paid'=>$newPaid_ammount,'due'=>$newDue_ammount]);
 
-        $delete_student_collection = DB::table('student_collection')->where('income_id',$id)->delete();
+            $delete_student_collection = DB::table('student_collection')->where('income_id',$id)->delete();
+        
+
+         
+        }
+
+
+
